backend/requirements.txt
fastapi
uvicorn
python-socketio
aiortc
python-dotenv
